+++ src/Repository/BigStorageRepository.php
@@ -2,6 +2,7 @@
 
 namespace Transip\Api\Library\Repository;
 
+use Psr\Http\Message\ResponseInterface;
 use Transip\Api\Library\Entity\BigStorage;
 
 class BigStorageRepository extends ApiRepository
@@ -58,7 +59,7 @@ class BigStorageRepository extends ApiRepository
         string $availabilityZone = '',
         string $vpsName = '',
         string $description = ''
-    ): void {
+    ): ResponseInterface {
         $parameters = [
             'size'             => $size,
             'offsiteBackups'   => $offsiteBackup,
@@ -66,7 +67,7 @@ class BigStorageRepository extends ApiRepository
             'vpsName'          => $vpsName,
             'description'      => $description,
         ];
-        $this->httpClient->post($this->getResourceUrl(), $parameters);
+        return $this->httpClient->post($this->getResourceUrl(), $parameters);
     }
 
     public function upgrade(string $bigStorageName, int $size, ?Bool $offsiteBackups = null): void
@@ -83,9 +84,9 @@ class BigStorageRepository extends ApiRepository
         $this->httpClient->post($this->getResourceUrl(), $parameters);
     }
 
-    public function update(BigStorage $bigStorage): void
+    public function update(BigStorage $bigStorage): ResponseInterface
     {
-        $this->httpClient->put(
+        return $this->httpClient->put(
             $this->getResourceUrl($bigStorage->getName()),
             ['bigStorage' => $bigStorage]
         );